pyproject.toml
[project]
name = "a2a_mcp"
version = "0.1.0"
description = "Agent-to-Agent vs Master Control Program communication patterns"
authors = [
    {name = "KhulnaSoft Lab", email = "dev.sulaiman@icloud.com"}
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.0",
    "requests>=2.20",
    "click>=8.0",
    "Flask-JWT-Extended>=4.5.2",
    "Flask-Limiter>=3.3.1",
    "marshmallow>=3.19.0",
    "tenacity>=8.2.2",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "PyJWT>=2.8.0",
    "prometheus-client>=0.17.0",
    "python-dateutil>=2.8.2",
    "gunicorn>=21.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
packages = ["src/a2a_mcp"]

[project.scripts]
a2a-mcp = "a2a_mcp.cli:cli" 